使っていない定数削除、マジックナンバー修正
