+++ googletest/tests/http_test.cpp
@@ -5,8 +5,7 @@
 #include <sys/types.h>
 
 #include "config/Config.hpp"
-#include "http/const/const_html_filename.hpp"
-#include "http/const/const_response_key_map.hpp"
+#include "http/const/const_header_field_values.hpp"
 #include "http/const/const_status_phrase.hpp"
 #include "http/request/RequestInfo.hpp"
 #include "http/response/Response.hpp"
