+++ srcs/http/const/const_header_field_values.hpp
@@ -1,0 +1,10 @@
+#ifndef SRCS_HTTP_CONST_CONST_RESPONSE_KEY_MAP_HPP
+#define SRCS_HTTP_CONST_CONST_RESPONSE_KEY_MAP_HPP
+
+#include <string>
+
+const std::string VERSION_HTTP     = "HTTP/1.1";
+const std::string CONNECTION_CLOSE = "close";
+const std::string TEXT_HTML        = "text/html";
+
+#endif /* SRCS_HTTP_CONST_CONST_RESPONSE_KEY_MAP_HPP */
