--- srcs/http/const/const_html_filename.hpp
@@ -1,9 +1,0 @@
-#ifndef SRCS_HTTP_CONST_CONST_HTML_FILENAME_HPP
-#define SRCS_HTTP_CONST_CONST_HTML_FILENAME_HPP
-
-#include <string>
-
-const std::string ROOT      = "/";
-const std::string TEXT_HTML = "text/html";
-
-#endif /* SRCS_HTTP_CONST_CONST_HTML_FILENAME_HPP */
