+++ srcs/http/response/Response.hpp
@@ -3,7 +3,7 @@
 
 #include "config/Config.hpp"
 #include "http/const/const_delimiter.hpp"
-#include "http/const/const_html_filename.hpp"
+#include "http/const/const_header_field_values.hpp"
 #include "http/const/const_status_phrase.hpp"
 #include "http/request/RequestInfo.hpp"
 #include "utils/utils.hpp"
@@ -12,7 +12,7 @@ class Response {
 private:
   const Config      &__config_;
   const RequestInfo &__request_info_;
-  int                __status_code_;
+  HttpStatusCode     __status_code_;
   std::string        __response_string_;
   std::string        __file_path_;
 
@@ -39,7 +39,7 @@ private:
 
   // clang-format off
   void __set_response_body() {
-    if (__status_code_ == 204 || __status_code_ == 304) {
+    if (__status_code_ == NO_CONTENT_204 || __status_code_ == 304) {
       return;
     }
     __set_content_len();
@@ -52,14 +52,14 @@ private:
     // start line
     __response_string_ = __version_ + SP + __status_phrase_ + CRLF;
     // header
-    if (!(__status_code_ == 204 || __status_code_ == 304)) {
+    if (!(__status_code_ == NO_CONTENT_204 || __status_code_ == 304)) {
       __response_string_ += "Content-Length: " + __content_len_ + CRLF;
       __response_string_ += "Content-Type: " + __content_type_ + CRLF;
     }
     __response_string_ += "Connection: " + __connection_ + CRLF;
     // empty line
     __response_string_ += CRLF;
-    if (__status_code_ == 204 || __status_code_ == 304) {
+    if (__status_code_ == NO_CONTENT_204 || __status_code_ == 304) {
       return;
     }
     // body
