+++ srcs/http/response/Response_delete.cpp
@@ -5,8 +5,7 @@
 #include <iostream>
 
 #include "config/Config.hpp"
-#include "http/const/const_html_filename.hpp"
-#include "http/const/const_response_key_map.hpp"
+#include "http/const/const_header_field_values.hpp"
 #include "http/const/const_status_phrase.hpp"
 #include "http/request/RequestInfo.hpp"
 #include "utils/file_io_utils.hpp"
